+++ apps/meteor/ee/client/omnichannel/additionalForms/MaxChatsPerAgentDisplay.tsx
@@ -2,9 +2,15 @@ import { useTranslation } from '@rocket.chat/ui-contexts';
 import React from 'react';
 
 import UserInfo from '../../../../client/components/UserInfo';
+import { useHasLicenseModule } from '../../hooks/useHasLicenseModule';
 
 const MaxChatsPerAgentDisplay = ({ maxNumberSimultaneousChat = 0 }) => {
 	const t = useTranslation();
+	const hasLicense = useHasLicenseModule('livechat-enterprise');
+
+	if (!hasLicense) {
+		return null;
+	}
 
 	return (
 		<>