+++ packages/livechat/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @rocket.chat/livechat Change Log
 
+## 1.14.5
+
+### Patch Changes
+
+- @rocket.chat/gazzodown@2.0.5
+
+## 1.14.4
+
+### Patch Changes
+
+- @rocket.chat/gazzodown@2.0.4
+
+## 1.14.3
+
+### Patch Changes
+
+- @rocket.chat/gazzodown@2.0.3
+
+## 1.14.2
+
+### Patch Changes
+
+- @rocket.chat/gazzodown@2.0.2
+
+## 1.14.1
+
+### Patch Changes
+
+- @rocket.chat/gazzodown@2.0.1
+
 ## 1.14.0
 
 ### Minor Changes