+++ packages/rest-typings/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @rocket.chat/rest-typings
 
+## 6.4.5
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.4.5
+
+## 6.4.4
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.4.4
+
+## 6.4.3
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.4.3
+
+## 6.4.2
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.4.2
+
+## 6.4.1
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.4.1
+
 ## 6.4.0
 
 ### Minor Changes